Tweak bargraph to select better channels.
Change the selected channels to pick ones closer to the wavelengths
of the 7 colours we're trying to detect.

This is still a suboptimal solution, since the measurements of these
channels go beyond the range of visible light and tend to saturate
easily.

Additionally there are many overlaps in the spectral response curves
that mean the actual intensity of any given wavelength is probably
more accurately measured as a function of two or more channels.